Add functionality to delete tag by name
If tag exists and there are addons wtih this tag , the records are delete first then the tag itself. If tag does not exist not found status is returned

+++ src/main/java/com/addonis/demo/services/TagServiceImpl.java
@@ -50,6 +50,11 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void deleteTagByName(String name) {
+        Tag tagToDelete = tagRepository.getTagByName(name);
+        if(tagToDelete == null) {
+            throw new EntityNotFoundException("tag", name);
+        }
+        tagRepository.deleteTagFromAllAddons(tagToDelete.getTagId());
         tagRepository.deleteTagByName(name);
     }
